Move messaging API calls to another file
