+++ src/components/Inbox/Inbox.js
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
+import { getAllUserConversations } from '../../utils/messagingAPICalls';
 import './Inbox.scss';
 import MessagePreview from '../MessagePreview/MessagePreview';
 
@@ -8,18 +9,7 @@ const Inbox = () => {
   const currentUser = useSelector(state => state.currentUser);
 
   const checkForMessages = () => {
-    const body = {"query": "{conversations(userId: \""+ currentUser.id + "\") {senderId recipientId}}"};
-
-    const options = {
-      method: 'POST',
-      body: JSON.stringify(body),
-      headers: {
-        'Content-Type': 'application/json',
-      }
-    };
-
-    fetch('https://ican2-be-rails.herokuapp.com/api/v1/graphql', options)
-      .then(response => response.json())
+    getAllUserConversations(currentUser.id)
       .then(data => {
         setConversations(data.data.conversations)
       })
